Port analysisrequests listing to senaite.core
