+++ src/senaite/core/browser/content/templates/samples.pt
